[REFACTOR] domain 패키지 내 enum 분리
