+++ src/main/java/com/oya/kr/popup/domain/enums/EntranceStatus.java
@@ -1,4 +1,4 @@
-package com.oya.kr.popup.domain;
+package com.oya.kr.popup.domain.enums;
 
 import static com.oya.kr.popup.exception.PlanErrorCodeList.NOT_EXIST_ENTRANCE_STATUS;
