+++ src/main/java/com/oya/kr/popup/domain/enums/WithdrawalStatus.java
@@ -1,4 +1,4 @@
-package com.oya.kr.popup.domain;
+package com.oya.kr.popup.domain.enums;
 
 import static com.oya.kr.popup.exception.PopupErrorCodeList.NOT_EXIST_WITHDRAWAL_STATUS;
