+++ src/main/java/com/oya/kr/popup/service/PlanService.java
@@ -20,8 +20,8 @@ import com.oya.kr.popup.controller.dto.response.DepartmentResponse;
 import com.oya.kr.popup.controller.dto.response.DepartmentsResponse;
 import com.oya.kr.popup.controller.dto.response.PlanResponse;
 import com.oya.kr.popup.controller.dto.response.PlansResponse;
-import com.oya.kr.popup.domain.Category;
-import com.oya.kr.popup.domain.Department;
+import com.oya.kr.popup.domain.enums.Category;
+import com.oya.kr.popup.domain.enums.Department;
 import com.oya.kr.popup.domain.Plan;
 import com.oya.kr.popup.mapper.PlanMapper;
 import com.oya.kr.popup.mapper.dto.request.PlanSaveMapperRequest;
