+++ src/test/java/com/oya/kr/popup/config/InitializingCategoryConfigTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import com.oya.kr.common.SpringApplicationTest;
-import com.oya.kr.popup.domain.Category;
+import com.oya.kr.popup.domain.enums.Category;
 import com.oya.kr.popup.mapper.CategoryMapper;
 import com.oya.kr.popup.mapper.dto.response.CategoryMapperResponse;
